requirements.txt
fastapi==0.109.0
pymongo==4.6.1
python-dotenv==1.0.0
uvicorn==0.25.0
python-multipart
pyjwt==2.8.0
authlib==1.3.0
itsdangerous==2.1.2